fix: Làm cho HealthBar và Score xuất hiện được

+++ include/structures/screens/gameplay/playing/score.h
@@ -8,9 +8,9 @@ namespace Structures::Screens::Gameplay::Playing::Score
 
 	struct Score : Collection
 	{
-		std::weak_ptr<HorizontalNumber<unsigned long>> score;
+		std::weak_ptr<HorizontalNumber<float>> score;
 		std::weak_ptr<HorizontalNumber<float>> accuracy;
 
-		Score(const TextureMemory& skin, const unsigned long* score, const float* accuracy);
+		Score(const TextureMemory& skin, const float* score, const float* accuracy);
 	};
 }
